[Fix #5247] Comment indentation before modifiers
Update `Layout/CommentIndentation`
to treat bare access modifiers like keywords
and allow a preceding comment to be more indented.

This fixes unexpected behavior in `Layout/CommentIndentation`
where comments were expected to be outdented
if they were followed by an outdented bare access modifier.

NOTE: This commit currently contains some commented code
for an alternative implementation of this fix.
The problem with the regex solution is that it duplicates
the listing of access modifiers elsewhere in the code.
The node solution does not look very good,
but I thought it was worth showing for consideration.